Added a cast to shut gcc up

+++ src/luav.c
@@ -101,6 +101,6 @@ void lv_dump(luav value) {
     case LNIL:        printf("nil\n");                      return;
   }
 
-  printf("Bad luav type: %lld\n", value & LUAV_TYPE_MASK);
+  printf("Bad luav type: %lld\n", (long long int) (value & LUAV_TYPE_MASK));
   assert(0 && "Bad luav type");
 }
